bots/requirements.txt
flask
APScheduler
Flask-APScheduler
# tweepy==3.10.0
tweepy
Flask-PyMongo
python-dotenv
gspread